Update versions of dask projects
This points to a consistent set of commits for dask/dask, dask/distributed,
and dask/dask-kubernetes.

In going through this I ran into a number of problems with Jupyter not starting
up well.  I tried to diagnose them by following github issues and pinned a few
extra versions of libraries like pyzmq, but in the end this was not effective.